requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
jinja2==3.1.2
gradio_client==0.8.1
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
